requirements.txt
Django==2.1.5
django-haystack==2.8.1
django-pure-pagination==0.3.0
django-ranged-response==0.2.0
django-simple-captcha==0.5.12
DjangoUeditor==2.0
jieba==0.42.1
Pillow==7.0.0
PyMySQL==0.9.3
pytz==2019.3
six==1.14.0
Whoosh==2.7.4
